+++ lib/src/launch_type.dart
@@ -1,6 +1,10 @@
 part of '../text_style_preview.dart';
 
+/// The type of launch for the [TextStylePreview].
 enum LaunchType {
+  /// The [TextStylePreview] is launched by tapping on the [Text].
   onTap,
+
+  /// The [TextStylePreview] is launched by long pressing on the [Text].
   onLongPress,
 }