+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+- Require Flutter sdk version changed to 3.0.0
+- Add reverse property.
+- Rename TextThemeType to ScaleCategory.
+
 ## 0.0.3
 
 - Fix typo.